refactor(structure): isolate complexity logic under lib/src/complexity
- Move all cognitive complexity visitor, analyzer, delta, git diff, and reporter implementations under lib/src/complexity/.

- Update public library exports in lib/cognitive_complexity.dart to point to new submodule paths.

- Update executable and test import references while maintaining 100% test pass and zero analyzer diagnostics.
