+++ lib/src/complexity/complexity_analyzer.dart
