+++ lib/src/complexity/delta_analyzer.dart
