+++ lib/src/complexity/git_diff_service.dart
